Updated libuv to 0.11.13 added scy::async namespace for thread primatives CMake build scope optimisations PacketStream now maintains an internal PacketQueue and async::Runner added deleter namespace to Memory.h for deleter functors

+++ src/HTTP/src/Message.cpp
@@ -86,7 +86,7 @@ UInt64 Message::getContentLength() const
 	{
 		return util::strtoi<UInt64>(contentLength);
 	}
-	else return UNKNOWN_CONTENT_LENGTH;
+	else return UInt64(UNKNOWN_CONTENT_LENGTH);
 }
